+++ Changelog.md
@@ -10,6 +10,7 @@
 - Change: Improved visuals when holding a taser
 - Change: Some alarm options have been moved to a separate screen
 - Change: The minimum radius of the Portable Radar is now one block
+- Change: The maximum allowed signature for a Keycard Reader is now 99999
 - API: Renamed Option#readFromNBT to Option#load and Option#writeToNBT to Option#save
 - Fix: Shields do not take damage when blocking a taser that deals high enough damage
 - Fix: The death message of a player killed by a taser does not contain the player firing the taser